fix(Route): Fix agency NPE issue
This fix is for feeds that have no agency records. When the agency_id for a route has a missing
reference, it is assigned the first agency if there is only one agency. If there are no agencies
present, the field is left null and a NoOperatorInFeedError is stored.

Fixes #40