bugfix release: cherry-pick PR #51
this allows us to load the grand-est GTFS feed.
Analysis depends on gtfs-lib 1.3, and we're hesitant to update it to v2 or v3 because we could corrupt the cached MapDBs.
Although if those are lazy-built, we could just delete the cached MapDBs and rebuild from original GTFS.
But for now, we'll just cherry-pick this patch so we can load this GTFS feed.