integrate graphql api directly into gtfs-lib
it used to be in a separate project called gtfs-api
we just put the Java graphql api definition in this project, which takes string queries and returns Object responses.
Whatever project includes gtfs-lib will need to provide an HTTP layer on top of this using some web server or framework.