feat(GraphQL): allow nested queries where fields don't  match
The GraphQL nested queries up until this change relied on parent tables with foreign refs sharing
the same field name as the primary key on the child table. For example, trips could be joined to a
route on the shared field route_id. This change permits nesting children stops under a parent stop
using stops#parent_station, which will be joined on the parent join value found in stops#stop_id.

refs conveyal/datatools-ui#428