fix(ReferenceTracker): properly track primary keys in proprietary files
Some proprietary tables in the GTFS+ spec do not conform to the general principle in GTFS where a
key field (e.g., stop_id) only acts as the primary key field in the entity's table. For example,
stop_id acts as a primary key on stop_attributes.txt, so we prepend the table name to the unique ID
for these tables when checking for duplicate entries.