create completely new file in dir with input GTFS, or reopen it.
The previous approach with the temp files was not working because the
temp file already exists and has size zero, so MapDB refuses to reopen.

Changing the definitions of the Maps to BTreeMaps or HTreeMaps I am not
having much luck. Loading seems far slower than MapDB1, which is what we
observed in previous testing. This could be due to some DB parameters.