fix(GTFSCache): close feeds when evicted, remove unused subclass.
The eviction listener was casting the FeedSource instances to GTFSFeed
to call close() on them, which was failing every time. I changed it to
only accept closeable types. But in fact we have only one type so we can
 eventually eliminate the generics and subclassing entirely.

Also temporarily evicting based on cache size, to try this out.
This will be less sensitive to GC cycles that reach near the max heap
size before clearing out a bunch of short-lived objects.