renamed methods, build but don't put, which is handled by cache
Previously we were calling methods that built networks but also
put them in the cache, even though they were being used as
value providers for a cache that stores its own values. This was
leading to a double-put, which caused a cache eviction every time
a new network was loaded. Once closing of MapDBs was working, this
caused the MapDB to be unreadable, because it is built, put, closed,
put again (automatically), and then read while closed.

also added stopgap reimplementation of getServiceDates.