Further refactoring of GTFSCache and GTFSFeed
GTFSFeed:
Abandon feedId, storing a specified unique ID directly in Atomic.String.

GTFSCache:
- Eliminate unused bucketFolder in GTFSCache.
- Factor out file name and File object creation.
- Factor out S3 file download method.
- Validate IDs instead of "cleaning" them.

Some of these changes are minimally invasive and improve readability,
but some break existing functionality, specifically the elimination of
the idGenerator function. The same methods are used when loading new
feeds or rebuilding old ones, and we have to exactly replicate the way
feedIds are assigned for existing old feeds.